requirements.txt
jsonschema==2.6.0
netaddr==0.7.19
pika==0.12.0
pyroute2==0.5.3
PyYAML==3.13
pyzmq==17.1.2
